refactor: refactor navbar
