+++ src/widgets/Navbar/ui/NavigationList.styled.ts
